requirements.txt
flask>=2.3.0
pillow>=10.0.0
google-genai>=0.5.0
werkzeug>=2.3.0
gunicorn>=21.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0